kits/DataFaker/pyproject.toml
[tool.poetry]
name = "datafaker"
version = "0.1.0"
description = ""
authors = ["DaiYuANg <dai_yuang@icloud.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9,<3.12"
faker = "^16.6.1"
pyside6 = "^6.4.2"
mysql-connector-python = "^8.0.32"
pypika = "^0.48.9"

[tool.poetry.dev-dependencies]
black = "^23.1.0"
pyinstaller = "^5.7.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

